requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.0
dataclasses-json==0.6.0
etelemetry==0.3.0
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.12.4
fitz==0.0.1.dev2
frozenlist==1.4.0
future==0.18.3
greenlet==2.0.2
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
Jinja2==3.1.2
langchain==0.0.297
langsmith==0.0.39
looseversion==1.3.0
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numexpr==2.8.6
numpy==1.26.0
openai==0.28.0
packaging==23.1
pandas==2.1.1
pathlib==1.0.1
Pillow==10.0.1
prov==2.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pydot==1.4.2
pyee==8.2.2
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pyparsing==3.1.1
PyPDF2==3.0.1
pyppeteer==1.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
pyxnat==1.6
PyYAML==6.0.1
rdflib==7.0.0
requests==2.31.0
scipy==1.11.2
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
starlette==0.27.0
tenacity==8.2.3
tqdm==4.66.1
traits==6.3.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
websockets==10.4
yarl==1.9.2
zipp==3.17.0
